New split screen mappings
I added a map for splitting the screen horrizontally, and and a
similar map for splitting the screen vertically. The move between
screens mappings were not working because of ^W instead of <C-W>.

I also added a map for moving up down left right with splits using
AWSD instead of HJKL since the leader is , and is on the right side
of the keyboard AWSD is on the left side of the keyboard.